requirements.txt
asgiref==3.2.10
boto3==1.14.38
botocore==1.17.38
Django==3.1
django-filter==2.3.0
django-materialize==1.0.1a2
django-materializecss-form==1.1.17
django-resized==0.3.11
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
